feat: pedersen & refractor ✨
