+++ primitives/starknet/Cargo.toml
@@ -13,6 +13,7 @@ targets = ["x86_64-unknown-linux-gnu"]
 [dependencies]
 starknet-crypto = {workspace = true,  default-features = false, features = ["alloc"]}
 starknet-ff = {workspace = true,  default-features = false, features = ["alloc"]}
+poseidon_hash = { default-features = false, package = "poseidon", git = "https://github.com/EvolveArt/poseidon-rs", branch = "feature/no-std-refractor", features = ["alloc"]}
 hex = { workspace = true, default-features = false, features = ["alloc"] }
 # Substrate
 frame-support = { workspace = true }
